+++ hotWaterSystemPIO/lib/Diverter/Diverter.h
@@ -6,16 +6,16 @@
 class Diverter
 {
 private:
-    int target; //= 90;       // target angle
-    int current; //= 90;      // current angle
+    int16_t target; //= 90;       // target angle
+    int16_t current; //= 90;      // current angle
     unsigned long interval = 20;      // delay time
     unsigned long previousMillis = 0;
 
 public:
     Servo servo;
 
-    void begin(int pin);
-    void setSpeed(int newSpeed);
+    void begin(uint8_t pin);
+    void setSpeed(unsigned long newSpeed);
     void open();
     void close();
     void update();
